requirements.txt
mne
numpy
scipy
pandas
matplotlib
torch
scikit-learn
pyyaml
omegaconf
tqdm
tensorboard
